Settings: Avoid enforcing weird backgrounds
Enforcing windowBackground on places they shouldn't be causes issues when applying background there.
Or another kind of issues, like full transparency on clear themes, or white bg on settings home categories.

Credits goes to Genkzsz11 and the whole Corvus-ROM team for finding some of the root issues on this commit:
https://github.com/Corvus-ROM/android_packages_apps_Settings/commit/2b595f3d5679bddfda6642bbdc55c3043d05dee1

Change-Id: Iacd5b2e369ba9089f5522ff637427b31dcdd69b7
Signed-off-by: S I P U N <sipunkumar85@gmail.com>